+++ ExpandWorldPrefabs/InfoSelector.cs
@@ -43,6 +43,8 @@ public class InfoSelector
     var waterY = pos.y - ZoneSystem.instance.m_waterLevel;
     var linq = data
       .Where(d => CheckArgs(d, args))
+      .Where(d => (d.Biomes & biome) == biome || d.AltBiomes != null)
+      .Where(d => (d.BannedBiomes & biome) == 0)
       .Where(d => (d.Biomes & biome.Biome) == biome.Biome)
       .Where(d => (d.BannedBiomes & biome.Biome) == 0)
       .Where(d => d.Day?.GetBool(f) != false || !day)
@@ -64,6 +66,11 @@ public class InfoSelector
       .Where(d => d.Condition == null || d.Condition.Evaluate(f));
     // Minor optimization to resolve simpler checks first (not measured).
     linq = [.. linq];
+    if (linq.Any(d => d.AltBiomes != null || d.BannedAltBiomes != null))
+    {
+      var altBiomes = WorldGenerator.instance.GetBiomeSector(pos).AltBiomes;
+      linq = [.. linq.Where(d => CheckBiomes(d, biome, altBiomes))];
+    }
     var checkEnvironments = linq.Any(d => d.Environments.Count > 0) || linq.Any(d => d.BannedEnvironments.Count > 0);
     var checkEvents = linq.Any(d => d.Events.Count > 0);
     var checkObjects = linq.Any(d => d.Objects != null);
@@ -85,7 +92,7 @@ public class InfoSelector
     }
     if (checkEnvironments)
     {
-      var environment = GetEnvironment(biome);
+      var environment = GetEnvironment(WorldGenerator.instance.GetBiomeSector(pos));
       linq = [.. linq
         .Where(d => d.Environments.Count == 0 || d.Environments.Contains(environment))
         .Where(d => !d.BannedEnvironments.Contains(environment))];
@@ -225,6 +232,11 @@ public class InfoSelector
     return true;
 
   }
+  internal static bool CheckBiomes(Info data, Heightmap.Biome biome, List<AltBiome> altBiomes) =>
+    ((data.Biomes & biome) == biome || altBiomes.Any(alt => data.AltBiomes?.Contains(alt.m_name) == true))
+    && (data.BannedBiomes & biome) == 0
+    && !altBiomes.Any(alt => data.BannedAltBiomes?.Contains(alt.m_name) == true);
+
   private static string GetEnvironment(BiomeSector biome)
   {
     var em = EnvMan.instance;